move models to separate modules
- move python.py and models.py to a models directory
- update unit tests
- use Dataset instead of ConjunctiveGraph per RDFLib/rdflib#2405
