+++ src/rdf_utils/models/common.py
